Commit 37

+++ pipelines/abalone/.ipynb_checkpoints/preprocess-checkpoint.py
@@ -75,7 +75,7 @@ if __name__ == "__main__":
         "medianIncome",
     ]
     ]
-    Y = df[["medianHouseValue"]] / 100000
+    Y = int(df[["medianHouseValue"]] / 100000)
 
     x_train, x_test, y_train, y_test = train_test_split(X, Y, test_size=0.33)
